telephony: Hack GSM and LTE signal strength
makeSignalStrengthFromRilParcel() is used for both solicited and
unsolicited requests, so we can hack it to send better signal strength
values to the modem backend. This significantly helps in areas where
GSM or LTE signal is poor but usable, in which case the modem would often
disconnect completely or fall back to a lower network mode (resulting in
poor UX).

Also, always report a valid ASU for LTE, even when the signal noise is
considered to be terrible, and adjust the threshold for terrible signal
noise in the signal icon as well.